Shader Work + Bubble
Created shader for the movement of seaweed.

Split Ripple shader into two parts, one for the screen effect and one for an object. Object version seems to have trouble with square objects though.

Tweaked the Bubble particle system. Made bubbles smaller, changes opacity of the sprite, and added random movement in the x-axis.